Moved to existing futures contract instrument
